feat: Shortcuts section with shared recorder row (#61)
