+++ Crisp/Resources/Localizable.xcstrings
@@ -1042,17 +1042,6 @@
         }
       }
     },
-    "HiDPI Shortcut" : {
-      "extractionState" : "stale",
-      "localizations" : {
-        "zh-Hans" : {
-          "stringUnit" : {
-            "state" : "translated",
-            "value" : "HiDPI 快捷键"
-          }
-        }
-      }
-    },
     "Highest" : {
       "localizations" : {
         "zh-Hans" : {
@@ -1569,7 +1558,6 @@
       }
     },
     "Record Shortcut" : {
-      "extractionState" : "stale",
       "localizations" : {
         "zh-Hans" : {
           "stringUnit" : {
@@ -1611,7 +1599,6 @@
       }
     },
     "Remove Shortcut" : {
-      "extractionState" : "stale",
       "localizations" : {
         "zh-Hans" : {
           "stringUnit" : {
@@ -1729,6 +1716,28 @@
           "stringUnit" : {
             "state" : "translated",
             "value" : "设置"
+          }
+        }
+      }
+    },
+    "Shortcut" : {
+      "extractionState" : "stale",
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "快捷键"
+          }
+        }
+      }
+    },
+    "Shortcuts" : {
+      "extractionState" : "stale",
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "快捷键"
           }
         }
       }
@@ -1875,6 +1884,17 @@
         }
       }
     },
+    "Toggle HiDPI" : {
+      "extractionState" : "stale",
+      "localizations" : {
+        "zh-Hans" : {
+          "stringUnit" : {
+            "state" : "translated",
+            "value" : "切换 HiDPI"
+          }
+        }
+      }
+    },
     "Tools" : {
       "extractionState" : "stale",
       "localizations" : {
@@ -1898,7 +1918,6 @@
       }
     },
     "Type shortcut… (Esc to cancel)" : {
-      "extractionState" : "stale",
       "localizations" : {
         "zh-Hans" : {
           "stringUnit" : {
